pyproject.toml
[tool.poetry]
name = "movie-recommendation-agent"
version = "0.1.0"
description = "A movie recommendation system using pydanticAI and TMDB API"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9"
pydantic-ai = "^0.0.10"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
jinja2 = "^3.1.3"
python-dotenv = "^1.0.1"
httpx = ">=0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
